% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{add_listfunc}
\alias{add_listfunc}
\title{Apply a function to a dataset creating another element in a list}
\usage{
add_listfunc(df, func, character = "Function")
}
\arguments{
\item{df}{A dataframe.}

\item{func}{A character string with the function to apply.}

\item{character}{A character string for the name of the element added to the list}
}
\value{
list
}
\description{
This function allows to create a list with the original dataset and a dataframe
containing the values originated by the application of the second argument to the dataset.
}
\examples{
dataset<-data.frame()
dataset_lsum <- add_listfunc(dataset, sum, 'TOT')
}
