% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{data_statplotly}
\alias{data_statplotly}
\title{Title}
\usage{
data_statplotly(
  autoc_dataset,
  main = "Moments of a Time series",
  yname = "Values",
  xname = "lags"
)
}
\arguments{
\item{xname}{}
}
\value{

}
\description{
Title
}
