% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{data_tracesl}
\alias{data_tracesl}
\title{Title}
\usage{
data_tracesl(
  dataset,
  n_ts1,
  n_ts2,
  main,
  second,
  title = "Comparison Time series plot",
  yname = "N° of sales",
  xname = "Dates"
)
}
\arguments{
\item{xname}{}
}
\value{

}
\description{
Title
}
