% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{data_tsplotly}
\alias{data_tsplotly}
\title{Title}
\usage{
data_tsplotly(
  dataset,
  n_ts,
  main = "Time series plot",
  yname = "N° of sales",
  xname = "Dates"
)
}
\arguments{
\item{xname}{}
}
\value{

}
\description{
Title
}
