% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dataset}
\alias{dataset}
\title{Sales of Testla Motorcars in the top-5 European Countries during a period from the third quarter
of 2013 to the fourth quarter of 2019.}
\format{A data frame with 26 rows and 5 variables}
\usage{
dataset
}
\description{
A dataset that contains real value-data for statistical porpuses with the table data can be found at this url.link:
https://teslamotorsc2013-0-01lub.com/tmc/threads/tesla-europe-registration-stats.61651/.
}
\keyword{datasets}
