% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{import_data}
\alias{import_data}
\title{Import data and filter the original dataset from "Tesla Europe Registration Stats"}
\usage{
import_data()
}
\value{
dataframe
}
\description{
This function grabs the dataset of the sales of Tesla in Europe form 2013 to 2019 and store the data frame to Global Environment.
Data taken from: https://teslamotorsclub.com/tmc/threads/tesla-europe-registration-stats.61651/.
Run this first function to recall the dataset.
}
\examples{
import_data()
}
