% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stat_data}
\alias{stat_data}
\title{To verify the order of integration of a time series and its statistics}
\usage{
stat_data(ts, n_ts)
}
\arguments{
\item{ts}{A class 'ts' object.}

\item{n_ts}{A number referring to the colum of the time series.}
}
\value{
A Charachter string and a class 'htest' object.
}
\description{
This function allows to verify if the time series is stationary or it is
integrated of order 1 or higher than 1.
}
\examples{
INSERT EXAMPLE
}
