% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stationary_data}
\alias{stationary_data}
\title{To verify the order of integration of a time series and its statistics}
\usage{
stationary_data(ts)
}
\arguments{
\item{ts}{A univariate time series of class 'ts' object.}
}
\value{
A Charachter string and a class 'htest' object.
}
\description{
This function allows to verify if the time series is stationary or it is
integrated of order 1 or higher than 1.
}
\examples{
dataset<-data.frame(...)
dataset_ts<-ts(data.frame,...)
stationary_data(dataset[,1])
}
