% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ts_data}
\alias{ts_data}
\title{Create a time series of a dataframe with the R-package zoo}
\usage{
ts_data(dataset, n_ts)
}
\arguments{
\item{dataset}{A dataframe.}

\item{n_ts}{Number of the column of the dataframe which will generate a time series.}
}
\value{
A class zoo object.
}
\description{
Create a time series of a dataframe with the R-package zoo
}
\examples{
dataset<-data.frame()
ts_data(dataset,1)
}
