% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ts_datastats}
\alias{ts_datastats}
\title{To create a time series of a dataframe}
\usage{
ts_datastats(dataset, start_date, end_date)
}
\arguments{
\item{dataset}{A dataframe.}

\item{start_date}{A numeric vector of two elemets, year and quarter, respectively.}

\item{end_date}{A numeric vector of two elemets, year and quarter, respectively.}
}
\value{
A class 'ts' object.
}
\description{
This function allows to create a time series object from a dataframe using
the R-package 'stats'.
}
\examples{
INSERT EXAMPLE
}
