% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ts_datazoo}
\alias{ts_datazoo}
\title{Create a time series of a dataframe}
\usage{
ts_datazoo(dataset = NULL)
}
\arguments{
\item{dataset}{A dataframe (default= 'NULL' which will recall the original dataset).}
}
\value{
A class 'zoo' object.
}
\description{
This function creates a time series of the dataset with the R-package 'zoo'
with the rownames used as dates (quarterly).
}
\examples{
dataset<-data.frame()
ts_data(dataset)
}
