% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{tsforc_data}
\alias{tsforc_data}
\title{To plot the linear forecast of an ARIMA model of a time series}
\usage{
tsforc_data(ts_arima, n_period, vctconf_interv, title, xname, yname)
}
\arguments{
\item{ts_arima}{A class 'Arima' object.}

\item{n_period}{A number refferred to the period of the forecast expressed in quarters.}

\item{vctconf_interv}{A numeric vector with two elements which refer respectively to the lower and upper
limits of the confidence interval of the forecast.}

\item{title}{A character string for the title of the plot.}

\item{xname}{A character string for the label of the x-axe.}

\item{yname}{A character string for the label of the y-axe.}
}
\value{
The plot of the forecast.
}
\description{
This function allows to create a plot which contains a linear forecast,
based on pre-selected ARIMA model.
}
\examples{
INSERT EXAMPLES
}
