% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{autoparam_tsarima}
\alias{autoparam_tsarima}
\title{To select automatically the best ARIMA model for a specified time series and
computing the corrispondent stationary time series}
\usage{
autoparam_tsarima(ts)
}
\arguments{
\item{ts}{A univariate time series of class 'ts' object.}
}
\value{
A list object composed by an object of class 'forecast_ARIMA' , "ARIMA" and 'Arima' (first element)
and an object of class "ts" (second element).
}
\description{
This function allows to look for the best parameters for the
tailored ARIMA model for a specific time series and computing the resulting stationary ts.
}
\examples{
dataset<-data.frame('something'=c(...))
dataset_ts<-ts(dataset, ...)
autoparam_tsarima(dataset_ts[,1])
}
