% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{barplotly_func}
\alias{barplotly_func}
\title{Create a barplot of a function applied to a dataframe}
\usage{
barplotly_func(
  dataset = NULL,
  func,
  main = "Barplot",
  xname = "Countries",
  yname = "Number of sales"
)
}
\arguments{
\item{dataset}{A dataframe (default= 'NULL' which will recall the original dataset).}

\item{func}{A function to be applied to the columns of the dataset.}

\item{main}{A character string for the name of the plot (default = 'Barplot').}

\item{xname}{A character string for the name of the x-axe (default = 'Countries').}

\item{yname}{A character string for the name of the y-axe (default = 'Number of sales').}
}
\value{
The plot for the selected arguments.
}
\description{
This function allows to crate a coloured barplot using the R-package plotly
of a function applied to to a specific dataframe.
}
\examples{
data_barplotly(func=sum, title='Barplot of dataset', xname='X-axe', yname='Y-axe')
}
