% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{data_barplotly}
\alias{data_barplotly}
\title{Create a barplot of a function applied to a dataframe}
\usage{
data_barplotly(dataset, func, main, xname, yname)
}
\arguments{
\item{dataset}{A dataframe.}

\item{func}{A function to be applied to the columns of the dataset.}

\item{main}{A character string for the name of the plot (default = 'Barplot').}

\item{xname}{A character string for the name of the x-axe (default = 'Countries').}

\item{yname}{A character string for the name of the y-axe (default = 'Number of sales').}
}
\value{
The plot for the selected arguments.
}
\description{
This function allows to crate a coloured barplot using the R-package plotly
of a function applied to to a specific dataframe.
}
\examples{
dataset<-add_listfunction(df,func,character)
data_barplotly(dataset, 'Barplot of dataset', 'X-axe', 'Y-axe')
}
