% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{data_barplotly}
\alias{data_barplotly}
\title{Create a barplot of the dataframe}
\usage{
data_barplotly(
  dataset,
  main = "Barplot",
  xname = "Countries",
  yname = "Number of sales"
)
}
\arguments{
\item{dataset}{A dataframe resulting by the function "add_listfunc".}

\item{main}{A character string for the name of the plot (default = 'Barplot').}

\item{xname}{A character string for the name of the x-axe (default = 'Countries')}

\item{yname}{A character string for the name of the y-axe (default = 'N° of sales')}
}
\value{
A barplot for the given arguments.
}
\description{
This function allows to crate a coloured barplot using the R-package plotly
of the second element of the list deriving by the add_listfunc().
}
\examples{
dataset<-add_listfunction(df,func,character)
data_barplotly(dataset, 'Barplot of dataset', 'X-axe', 'Y-axe')
}
