% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{data_tracesl}
\alias{data_tracesl}
\title{Plot two time series from a dataset in the same linegraph}
\usage{
data_tracesl(
  dataset1,
  dataset2,
  start_date = "2013-09-30",
  n_obs = 26,
  interval = "quarter",
  main,
  second,
  title = "Comparison Time series plot",
  yname = "Number of sales",
  xname = "Dates"
)
}
\arguments{
\item{dataset1}{A subsetted dataframe which contains one precise time series of interest
(to allow for univariate ts).}

\item{dataset2}{A subsetted dataframe which contains one precise time series of interest
(to allow for univariate ts).}

\item{start_date}{A character string which refers to the date of the first observation, with format "%Y-%m-%d" (default = "2013-09-30").}

\item{n_obs}{A number equal to the observations inside the time series (default = 26).}

\item{interval}{A character string which refers to the time interval between each observation (default = "quarter").}

\item{main}{A character string for the name of the first time series.}

\item{second}{A character string for the name of the second time series.}

\item{title}{A character string for the title of the plot (default = 'Comparison Time series plot').}

\item{yname}{A character string for the name of the y-axe (default = 'Number of sales').}

\item{xname}{A character string for the name of the x-axe (default = 'Dates').}
}
\value{
The plot for the selected arguments.
}
\description{
This function allows comparisons in trends between two timeseries selected
from a dataset and includes a slider to better select the periods of the analysis.
}
\examples{
dataset<-data.frame(..., 'something'=c(...), 'somethingelse'=c(...))
data_tracesl(dataset$something,dataset$somethingelse,main='High school', main='University', second='High school',title='Grades', yname='Numbers')
}
