% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{data_tsplotly}
\alias{data_tsplotly}
\title{Create a lineplot of a time series}
\usage{
data_tsplotly(
  dataset,
  n_ts,
  main = "Time series plot",
  yname = "Number of sales",
  xname = "Dates"
)
}
\arguments{
\item{dataset}{A dataframe.}

\item{n_ts}{A number referring to the colum of the dataset to plot as a time series.}

\item{main}{A character string for the title of the plot (default = 'Time series plot')}

\item{yname}{A character string for the name of the x-axe (default = 'Dates')}

\item{xname}{A character string for the name of the y-axe (default = 'Number of sales')}
}
\value{
The plot for the selected arguments.
}
\description{
This function creates a lines+scatter plot of a time series using
the R-package 'plotly'.
}
\examples{
dataset<-data.frame()
data_tsplotly(dataset, 1, 'Graph', 'Values', 'Dates')
}
