% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{manualparam_tsarima}
\alias{manualparam_tsarima}
\title{To select manually the best ARIMA model for a specified time series and computing the
correspoding stationary time series}
\usage{
manualparam_tsarima(ts, vct_param)
}
\arguments{
\item{ts}{A univariate time series of class 'ts' object.}

\item{vct_param}{A numeric vector with three integer components (p, d, q), which refers to
the AR order, the degree of differencing, and the MA order.}
}
\value{
A list object composed by an object of class 'forecast_ARIMA' , "ARIMA" and 'Arima' (first element)
and an object of class "ts" (second element).
}
\description{
This function allows to manually look for the best parameters for the
tailored ARIMA model for a specific time series and computing the resulting stationary ts.
}
\examples{
dataset<-data.frame('something'=c(...))
dataset_ts<-ts(dataset, ...)
manualparam_tsarima(dataset_ts[,1])
}
