% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{manualparam_tsarima}
\alias{manualparam_tsarima}
\title{To select manually the best ARIMA model for a specified time series.}
\usage{
manualparam_tsarima(ts, n_ts, vct_param)
}
\arguments{
\item{ts}{A class 'ts' object.}

\item{n_ts}{A number referring to the colum of the time series.}

\item{vct_param}{A numeric vector with three integer components (p, d, q), which refers to
the AR order, the degree of differencing, and the MA order.}
}
\value{
A class 'Arima' object.
}
\description{
To select manually the best ARIMA model for a specified time series.
}
