% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_tsres}
\alias{plot_tsres}
\title{To compute and plot the residualsa and the ACF of an ARIMA model of a Time Series}
\usage{
plot_tsres(ts_arima, n_degfree)
}
\arguments{
\item{ts_arima}{A univariate time series with class 'Arima' object.}

\item{n_degfree}{A number referred to the degrees of freedom for the residuals.}
}
\value{
Three plots (Residuals, ACF and count) or a character string referred to an argument error.
}
\description{
This function allows to create three separate plots of the residuals,
Partial autocorrelation (PAC) anc the count of an ARIMA model of a time series.
}
\examples{
dataset<-data.frame('something'=c(...))
dataset_ts<-ts(dataset, ...)
arima_ts<-autoparam_tsarima(dataset_ts[,1])
plot_tsres(arima_ts, "ARIMA")
}
