% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_tsres}
\alias{plot_tsres}
\title{To plot the residuals of a ARIMA model of a Time Series}
\usage{
plot_tsres(ts_arima, title = "PLOT ARIMA MODEL")
}
\arguments{
\item{ts_arima}{A univariate time series with class 'Arima' object.}

\item{title}{A character string for the title of the plot.}
}
\value{
Three plots (Residuals, ACF and PACF).
}
\description{
This function allows to create a plot of the residuals of an ARIMA model
of a time series with both two separate plots of the Autocorrelation (ACF) and Partial Autocorrelation function (PACF)
}
\examples{
dataset<-data.frame(...)
dataset_ts<-ts(data.frame,...)
arima_ts<-autoparam_tsarima(dataset_ts[,1])
plot_tsres(arima_ts, "ARIMA")
}
