% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ts_datastats}
\alias{ts_datastats}
\title{To create a time series of a dataframe}
\usage{
ts_datastats(dataset = NULL, start_date = c(2013, 3), end_date = c(2019, 4))
}
\arguments{
\item{dataset}{A dataframe.}

\item{start_date}{A numeric vector of two elemets, year and quarter, respectively.}

\item{end_date}{A numeric vector of two elemets, year and quarter, respectively.}
}
\value{
A class 'ts' object.
}
\description{
This function allows to create a time series object from a dataframe using
the R-package 'stats'.
}
\examples{
dataset<-data.frame(...)
data_ts(dataset,start_date=c(2010,2),end_data=c(2020,4))
}
