% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ts_datazoo}
\alias{ts_datazoo}
\title{Create a time series of a dataframe}
\usage{
ts_datazoo(dataset, n_ts)
}
\arguments{
\item{dataset}{A dataframe.}

\item{n_ts}{Number of the column of the dataframe which will generate a time series.}
}
\value{
A class 'zoo' object.
}
\description{
This function creates a time series of the dataset with teh R-package 'zoo'
with the rownames used as Dates (quarterly).
}
\examples{
dataset<-data.frame()
ts_data(dataset,1)
}
